+++ build.sh
@@ -1,6 +1,12 @@
 #/bash/sh
 export GOPROXY=direct
 
+# nps embeds web/dist via go:embed; build the admin UI first.
+if [ ! -f web/dist/index.html ]; then
+  echo "web/dist/index.html missing - build the admin UI first (cd web/ui && pnpm install && pnpm build)" >&2
+  exit 1
+fi
+
 sudo apt-get update
 sudo apt-get install gcc-mingw-w64-i686 gcc-multilib
 env GOOS=windows GOARCH=386 CGO_ENABLED=1 CC=i686-w64-mingw32-gcc go build -ldflags "-s -w -extldflags -static -extldflags -static" -buildmode=c-shared -o npc_sdk.dll cmd/npc/sdk.go
@@ -80,71 +86,71 @@ tar -czvf darwin_arm64_client.tar.gz npc conf/npc.conf conf/multi_account.conf
 
 
 CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -ldflags "-s -w -extldflags -static -extldflags -static" ./cmd/nps/nps.go
-tar -czvf linux_amd64_server.tar.gz conf/nps.conf web/views web/static nps
+tar -czvf linux_amd64_server.tar.gz conf/nps.conf nps
 
 
 CGO_ENABLED=0 GOOS=linux GOARCH=386 go build -ldflags "-s -w -extldflags -static -extldflags -static" ./cmd/nps/nps.go
-tar -czvf linux_386_server.tar.gz conf/nps.conf web/views web/static nps
+tar -czvf linux_386_server.tar.gz conf/nps.conf nps
 
 
 CGO_ENABLED=0 GOOS=linux GOARCH=arm GOARM=5 go build -ldflags "-s -w -extldflags -static -extldflags -static" ./cmd/nps/nps.go
-tar -czvf linux_arm_v5_server.tar.gz conf/nps.conf web/views web/static nps
+tar -czvf linux_arm_v5_server.tar.gz conf/nps.conf nps
 
 
 CGO_ENABLED=0 GOOS=linux GOARCH=arm GOARM=6 go build -ldflags "-s -w -extldflags -static -extldflags -static" ./cmd/nps/nps.go
-tar -czvf linux_arm_v6_server.tar.gz conf/nps.conf web/views web/static nps
+tar -czvf linux_arm_v6_server.tar.gz conf/nps.conf nps
 
 
 CGO_ENABLED=0 GOOS=linux GOARCH=arm GOARM=7 go build -ldflags "-s -w -extldflags -static -extldflags -static" ./cmd/nps/nps.go
-tar -czvf linux_arm_v7_server.tar.gz conf/nps.conf web/views web/static nps
+tar -czvf linux_arm_v7_server.tar.gz conf/nps.conf nps
 
 
 CGO_ENABLED=0 GOOS=linux GOARCH=arm64 go build -ldflags "-s -w -extldflags -static -extldflags -static" ./cmd/nps/nps.go
-tar -czvf linux_arm64_server.tar.gz conf/nps.conf web/views web/static nps
+tar -czvf linux_arm64_server.tar.gz conf/nps.conf nps
 
 
 CGO_ENABLED=0 GOOS=freebsd GOARCH=arm go build -ldflags "-s -w -extldflags -static -extldflags -static" ./cmd/nps/nps.go
-tar -czvf freebsd_arm_server.tar.gz conf/nps.conf web/views web/static nps
+tar -czvf freebsd_arm_server.tar.gz conf/nps.conf nps
 
 
 CGO_ENABLED=0 GOOS=freebsd GOARCH=386 go build -ldflags "-s -w -extldflags -static -extldflags -static" ./cmd/nps/nps.go
-tar -czvf freebsd_386_server.tar.gz conf/nps.conf web/views web/static nps
+tar -czvf freebsd_386_server.tar.gz conf/nps.conf nps
 
 
 CGO_ENABLED=0 GOOS=freebsd GOARCH=amd64 go build -ldflags "-s -w -extldflags -static -extldflags -static" ./cmd/nps/nps.go
-tar -czvf freebsd_amd64_server.tar.gz conf/nps.conf web/views web/static nps
+tar -czvf freebsd_amd64_server.tar.gz conf/nps.conf nps
 
 
 CGO_ENABLED=0 GOOS=linux GOARCH=mips go build -ldflags "-s -w -extldflags -static -extldflags -static" ./cmd/nps/nps.go
-tar -czvf linux_mips_server.tar.gz conf/nps.conf web/views web/static nps
+tar -czvf linux_mips_server.tar.gz conf/nps.conf nps
 
 
 CGO_ENABLED=0 GOOS=linux GOARCH=mips64 go build -ldflags "-s -w -extldflags -static -extldflags -static" ./cmd/nps/nps.go
-tar -czvf linux_mips64_server.tar.gz conf/nps.conf web/views web/static nps
+tar -czvf linux_mips64_server.tar.gz conf/nps.conf nps
 
 
 CGO_ENABLED=0 GOOS=linux GOARCH=mips64le go build -ldflags "-s -w -extldflags -static -extldflags -static" ./cmd/nps/nps.go
-tar -czvf linux_mips64le_server.tar.gz conf/nps.conf web/views web/static nps
+tar -czvf linux_mips64le_server.tar.gz conf/nps.conf nps
 
 
 CGO_ENABLED=0 GOOS=linux GOARCH=mipsle go build -ldflags "-s -w -extldflags -static -extldflags -static" ./cmd/nps/nps.go
-tar -czvf linux_mipsle_server.tar.gz conf/nps.conf web/views web/static nps
+tar -czvf linux_mipsle_server.tar.gz conf/nps.conf nps
 
 
 CGO_ENABLED=0 GOOS=darwin GOARCH=amd64 go build -ldflags "-s -w -extldflags -static -extldflags -static" ./cmd/nps/nps.go
-tar -czvf darwin_amd64_server.tar.gz conf/nps.conf web/views web/static nps
+tar -czvf darwin_amd64_server.tar.gz conf/nps.conf nps
 
 
 CGO_ENABLED=0 GOOS=darwin GOARCH=arm64 go build -ldflags "-s -w -extldflags -static -extldflags -static" ./cmd/nps/nps.go
-tar -czvf darwin_arm64_server.tar.gz conf/nps.conf web/views web/static nps
+tar -czvf darwin_arm64_server.tar.gz conf/nps.conf nps
 
 
 CGO_ENABLED=0 GOOS=windows GOARCH=amd64 go build -ldflags "-s -w -extldflags -static -extldflags -static" ./cmd/nps/nps.go
-tar -czvf windows_amd64_server.tar.gz conf/nps.conf web/views web/static nps.exe
+tar -czvf windows_amd64_server.tar.gz conf/nps.conf nps.exe
 
 
 CGO_ENABLED=0 GOOS=windows GOARCH=386 go build -ldflags "-s -w -extldflags -static -extldflags -static" ./cmd/nps/nps.go
-tar -czvf windows_386_server.tar.gz conf/nps.conf web/views web/static nps.exe
+tar -czvf windows_386_server.tar.gz conf/nps.conf nps.exe
 
 VERSION=`sed -n '/VERSION/s/.*"\(.*\)"/\1/p' lib/version/version.go`
